professionals/ai/requirements.txt
-e ../../pkgs/py-pkg
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
sse-starlette>=2.1.0
google-genai>=1.0.0
python-jose[cryptography]>=3.3.0
structlog>=24.4.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
